LPD-8428 Move fixture and page inside tests folder
